Added passwordeval option and documentation.
Removes need for plain text password in the getmailrc file. Instead of "password"
option "passwordeval" option can be used.

For example:

passwordeval = gpg -d gmail.gpg

The value of the "passwordeval" option is run and the output is used as the
password. If password cannot be found from the output, then password is prompted
from the user.